Consolidate account page into profile page
* Move classroom management to profile page
* Move category/score-over-time progression charts to top of problems
  (TODO: convert from underscore.js template to react componetns)
* Remove use of react-addon update for setState $set operations
  (TODO: phase out $merge and $push remainders)
* Clean up unused underscorejs templates for problem page
* Clean up some react-bootstrap imports
  (TODO: go through rest, sort and remove unused)
