+++ picoCTF-web/web/_includes/header.html
@@ -41,12 +41,8 @@
             <ul class="nav navbar-nav navbar-right">
                 <% if(status.logged_in) { %>
                 <li>
-                    <a href="/account" id="navbar-item-account">
-                        <span class="glyphicon glyphicon-user" aria-hidden="true"></span> <%- status.username %>
-                    </a>
-                </li>
-                <li>
-                    <a href="/profile" id="navbar-item-profile">
+                    <a href="/profile" id="navbar-item-account">
+                        <span class="glyphicon glyphicon-user" aria-hidden="true"></span> <%- status.username %>&nbsp;&nbsp;
                         <span class="glyphicon glyphicon-flash" aria-hidden="true"></span> <%- status.score %>
                     </a>
                 </li>
