--- picoCTF-web/web/account.html
@@ -1,85 +1,0 @@
----
-layout: default
-title: Account
-post_scripts:
- - js/account.js
- - js/libs/download.js
-startup_functions:
- - redirectIfNotLoggedIn()
----
-
-<div class="container">
-  <div class="row">
-    <div class="col-md-6" id="team-management-container" style="display:none">
-      <div id="team-management"></div>
-    </div>
-    <div class="col-md-6">
-      <div class="panel panel-default">
-        <div class="panel-heading">
-          <h3 class="panel-title">Update User Password</h3>
-        </div>
-        <div class="panel-body">
-          <div class="col-md-10 col-md-offset-1">
-            <form role="form" id="password-update-form">
-              <div class="form-group">
-                <label for="current-password" class="control-label">Current Password</label>
-                <input name="current-password" id="current-password" class="form-control" type="password" placeholder="Current Password">
-              </div>
-              <div class="form-group">
-                <label for="new-password" class="control-label">New Password</label>
-                <input name="new-password" id="new-password" class="form-control" type="password" placeholder="New Password">
-              </div>
-              <div class="form-group">
-                <label for="new-password-confirmation" class="control-label">Repeat New Password</label>
-                <input name="new-password-confirmation" id="new-password-confirmation" class="form-control" type="password" placeholder="Repeat New Password">
-              </div>
-              <div class="form-group">
-                <button id="password-update-button" class="btn btn-primary" type="submit">Update Password</button>
-              </div>
-            </form>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="col-md-6">
-      <div class="panel panel-default">
-        <div class="panel-heading">
-          <h3 class="panel-title">Download Account Data</h3>
-        </div>
-        <div class="panel-body">
-          <div class="col-md-10 col-md-offset-1">
-            <form role="form" id="download-data-form">
-                <p>Click here to download all data associated with your account.</p>
-              <div class="form-group">
-                <button id="download-data-button" class="btn btn-primary" type="submit">Download Account Data</button>
-              </div>
-            </form>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="col-md-6">
-      <div class="panel panel-danger">
-        <div class="panel-heading">
-          <h3 class="panel-title">Delete Account</h3>
-          <em>This is permanent action; you can not undo this.</em>
-        </div>
-        <div class="panel-body">
-          <div class="col-md-10 col-md-offset-1">
-            <form role="form" id="disable-account-form">
-              <p>Deleting your account will remove you from the competition. You will not be able to log in once you delete your account. Deleted accounts cannot be re-enabled.</p>
-              <div class="form-group">
-                <label for="current-password" class="control-label">Current Password</label>
-                <input name="current-password" id="current-password-disable" class="form-control" type="password" placeholder="Current Password">
-              </div>
-              <div class="form-group">
-                <button id="disable-account-button" class="btn btn-danger" type="submit">Delete Account</button>
-              </div>
-            </form>
-          </div>
-        </div>
-      </div>
-    </div>
-
-  </div>
-</div>
