+++ picoCTF-web/web/css/main.css
@@ -385,7 +385,7 @@ canvas.chartjs-render-monitor {
 }
 
 .form-inline .class-teacher-username {
-    width: 160px;
+    width: 210px;
 }
 
 /* Registration */
