+++ picoCTF-web/web/index.html
@@ -2,13 +2,14 @@
 layout: default
 title: Homepage
 title_label: hide
+pre_scripts:
+- /js/libs/jquery.bbq.js
 post_scripts:
 - /js/react-utils.js
 - /js/libs/react-recaptcha.js
 - /js/front-page.js
 ---
 
-<script src="/js/libs/jquery.bbq.js"></script>
 <script src="https://www.google.com/recaptcha/api.js?render=explicit" async defer></script>
 
 <div class="container">
