+++ picoCTF-web/web/problems.html
@@ -3,134 +3,14 @@ layout: default
 title: Problems
 post_scripts:
 - js/problems.js
+- js/progression_graphs.js
+- /js/libs/Chart.min.js
+- js/libs/star-rating.min.js
 startup_functions:
 - redirectIfNotLoggedIn()
 ---
 
-<script src="/js/libs/star-rating.min.js"></script>
 <link href="/css/star-rating.min.css" rel="stylesheet" type="text/css"/>
-
-<script type="text/template" id="problem-list-template">
-    <div class="problem-list panel-group">
-        <% _.each(problems, function(problem) { %>
-        <%= renderProblem({
-        problem: problem,
-        reviewData: reviewData,
-        renderProblemSubmit: renderProblemSubmit,
-        sanitizeMetricName: sanitizeMetricName
-        }) %>
-        <% }); %>
-    </div>
-</script>
-
-<script type="text/template" id="problem-template">
-    <%
-    var alreadyReviewed = false;
-    var review = null;
-
-    for (var i = 0; i < reviewData.length; i++)
-      if (reviewData[i].pid == problem.pid){
-        alreadyReviewed = true;
-        review = reviewData[i].feedback;
-      }
-
-
-    var thumbs = {
-      upClass: alreadyReviewed && review.liked ? "active" : "",
-      downClass: alreadyReviewed && !review.liked ? "active" : ""
-    };
-    var panel = {
-    className: problem.solved ? "solved" : "",
-    collapseClass: problem.solved ? "" : "in",
-    title: problem.name + " - Points: " + problem.score + " - (Solves: " + problem.solves + ")",
-    status: (problem.unlocked ? (problem.solved ? "Solved" : "Unsolved") : "Locked"),
-    };
-
-    var pill = {
-    solveId: problem.pid + "solve",
-    reviewId: problem.pid + "review",
-    hintId: problem.pid + "hint",
-    solveClassName: "active",
-    };
-    %>
-
-    <div class="problem <%= panel.className %> panel panel-default">
-        <div class="problem-header panel-heading" data-toggle="collapse" data-target="#<%- problem.pid %>">
-            <h4 class="panel-title">
-                <%- panel.title %>
-                <div class="pull-right"><%- problem.category %> - <%- panel.status %></div>
-            </h4>
-        </div>
-
-        <div class="panel-collapse collapse <%= panel.collapseClass %>" id="<%= problem.pid %>">
-            <% if (problem.unlocked) { %>
-              <div class="problem-body panel-body">
-                  <ul class="nav nav-tabs">
-                      <li class="<%= pill.solveClassName %>">
-                        <a href="#<%= pill.solveId %>" data-toggle="tab" class="solve-tab-button" data-pid="<%= problem.pid %>">Solve</a>
-                      </li>
-                      <li>
-                        <a href="#<%= pill.hintId %>" data-toggle="tab" class="hint-tab-button" data-pid="<%= problem.pid %>">Hints</a>
-                      </li>
-                  </ul>
-                  <div class="tab-content">
-                      <div class="tab-pane <%= pill.solveClassName %>" id="<%= pill.solveId %>">
-                          <div class="panel-body">
-                              <%= renderProblemSubmit({problem: problem, thumbs: thumbs}) %>
-                          </div>
-                      </div>
-                      <div class="tab-pane" id="<%= pill.hintId %>">
-                          <div class="panel-body">
-                              <ul class="list-group">
-                                <% for (var i = 0; i < problem.hints.length; i++){ %>
-                                  <li class="list-group-item"><%= problem.hints[i] %></li>
-                                <% } %>
-                              </ul>
-                          </div>
-                      </div>
-                  </div>
-              </div>
-            <% } %>
-        </div>
-    </div>
-</script>
-
-<script type="text/template" id="problem-submit-template">
-    <p class="problem-description">
-        <%= problem.description %>
-    <div class="row">
-        <div class="col-md-11">
-            <% if(problem.hint) { %>
-            <em class="problem-hint pull-left" id="<%= problem.pid %>-hint">
-                <strong><h4>Hint:</h4></strong>
-                <%= problem.hint %>
-            </em>
-            <% } %>
-        </div>
-        <div class="col-md-1">
-            <% if(problem.hint && false) { %>
-            <span data-pid="<%= problem.pid %>" class="glyphicon glyphicon-info-sign info-span pull-right"></span>
-            <% } %>
-        </div>
-    </div>
-    </p>
-    <div class="row">
-        <form class="problem-submit">
-            <div class="col-md-10 col-sm-9 col-xs-7 col-lg-11">
-              <div class="input-group input-group-sm">
-                <span class="input-group-btn">
-                  <button class="btn btn-primary" type="submit">Submit!</button>
-                </span>
-                <input type="text" data-pid="<%- problem.pid %>" class="form-control" placeholder="picoCTF{FLAG}"/>
-              </div>
-            </div>
-            <div class="col-md-2 col-sm-3 col-xs-5 col-lg-1">
-                <span id="<%- problem.pid %>-thumbsup" data-pid="<%- problem.pid %>" data-setting="up" style="font-size:1.7em;" class="rating-button glyphicon glyphicon-thumbs-up pull-left <%- thumbs.upClass %>"></span>
-                <span id="<%- problem.pid %>-thumbsdown" data-pid="<%- problem.pid %>" data-setting="down" style="font-size:1.7em;" class="rating-button glyphicon glyphicon-thumbs-down pull-right <%- thumbs.downClass %>"></span>
-            </div>
-        </form>
-    </div>
-</script>
 
 <script type="text/template" id="achievement-message-template">
     <table>
@@ -147,5 +27,40 @@ startup_functions:
 </script>
 
 <div class="container">
+  <div class="row">
+    <div class="panel-group">
+      <div class="panel panel-default">
+        <div class="panel-heading">
+          <h4 class="panel-title">
+            <a data-toggle="collapse" href="#progress-collapse">
+              <span class="glyphicon glyphicon-stats" aria-hidden="true"></span>
+              Progress Tracker</a>
+          </h4>
+        </div>
+        <div id="progress-collapse" class="panel-collapse collapse">
+          <div class="panel-body">
+            <div class="row">
+              <div class="col-md-6">
+                <div id="progress-info"></div>
+              </div>
+              <div class="col-md-6">
+                <div class="panel panel-default"
+                     id="team-progression-graph-container">
+                  <div class="panel-heading">
+                    <h3 class="panel-title">Score Progression over Time</h3>
+                  </div>
+                  <div class="panel-body team-panel-body">
+                    <div id="team-progression-graph"></div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="row">
     <div id="problem-list-holder"></div>
+  </div>
 </div>
