+++ picoCTF-web/web/profile.html
@@ -2,30 +2,80 @@
 layout: default
 title: Profile
 post_scripts:
- - js/progression_graphs.js
  - js/profile.js
+ - js/libs/download.js
 startup_functions:
  - redirectIfNotLoggedIn()
 ---
 
-<script type="text/javascript" src="/js/libs/Chart.min.js"></script>
-
 <div class="container">
   <div class="row">
-    <div class="col-md-6">
-      <div id="progress-info"></div>
-      <div class="panel panel-default" id="team-progression-graph-container" >
-          <div class="panel-heading">
-             <h3 class="panel-title">Score Progression</h3>
-          </div>
-          <div class="panel-body team-panel-body">
-            <div id="team-progression-graph"></div>
-          </div>
-      </div>
+    <div class="col-md-6" >
+      <div id="group-info"></div>
+      <div id="team-management"></div>
     </div>
     <div class="col-md-6">
-      <div id="group-info"></div>
-      <div id="achievement-info"></div>
+      <div class="panel panel-default">
+        <div class="panel-heading">
+          <h3 class="panel-title">Update User Password</h3>
+        </div>
+        <div class="panel-body">
+          <div class="col-md-10 col-md-offset-1">
+            <form role="form" id="password-update-form">
+              <div class="form-group">
+                <label for="current-password" class="control-label">Current Password</label>
+                <input name="current-password" id="current-password" class="form-control" type="password" placeholder="Current Password">
+              </div>
+              <div class="form-group">
+                <label for="new-password" class="control-label">New Password</label>
+                <input name="new-password" id="new-password" class="form-control" type="password" placeholder="New Password">
+              </div>
+              <div class="form-group">
+                <label for="new-password-confirmation" class="control-label">Repeat New Password</label>
+                <input name="new-password-confirmation" id="new-password-confirmation" class="form-control" type="password" placeholder="Repeat New Password">
+              </div>
+              <div class="form-group">
+                <button id="password-update-button" class="btn btn-primary" type="submit">Update Password</button>
+              </div>
+            </form>
+          </div>
+        </div>
+      </div>
+      <div class="panel panel-default">
+        <div class="panel-heading">
+          <h3 class="panel-title">Download Account Data</h3>
+        </div>
+        <div class="panel-body">
+          <div class="col-md-10 col-md-offset-1">
+            <form role="form" id="download-data-form">
+                <p>Click here to download all data associated with your account.</p>
+              <div class="form-group">
+                <button id="download-data-button" class="btn btn-primary" type="submit">Download Account Data</button>
+              </div>
+            </form>
+          </div>
+        </div>
+      </div>
+      <div class="panel panel-danger">
+        <div class="panel-heading">
+          <h3 class="panel-title">Delete Account</h3>
+          <em>This is permanent action; you can not undo this.</em>
+        </div>
+        <div class="panel-body">
+          <div class="col-md-10 col-md-offset-1">
+            <form role="form" id="disable-account-form">
+              <p>Deleting your account will remove you from the competition. You will not be able to log in once you delete your account. Deleted accounts cannot be re-enabled.</p>
+              <div class="form-group">
+                <label for="current-password" class="control-label">Current Password</label>
+                <input name="current-password" id="current-password-disable" class="form-control" type="password" placeholder="Current Password">
+              </div>
+              <div class="form-group">
+                <button id="disable-account-button" class="btn btn-danger" type="submit">Delete Account</button>
+              </div>
+            </form>
+          </div>
+        </div>
+      </div>
     </div>
   </div>
 </div>
@@ -74,68 +124,7 @@ startup_functions:
             <input type="text" class="form-control class-teacher-username" id="group-owner-input" placeholder="Classroom Owner Username">
         </div>
         <button class="btn btn-primary" id="join-group">Join</button>
-        </div>
     </form>
     </div>
-</script>
-
-<script type="text/template" id="team-info-template">
-  <div class="panel panel-default">
-    <div class="panel-heading">
-      <h3 class="panel-title clearfix"><b><%- data.team_name %></b><div class="pull-right"><%- "Score: "+data.score %></strong></h3>
-    </div>
-    <div class="panel-body">
-      <p>Your Team Passphrase is <strong><%- data.password %></strong>.</p>
-      <p>Members:</p>
-      <ul>
-        <% _.each(data.members, function(username) { %>
-          <li><%- username %></li>
-        <% }); %>
-      </ul>
-        <% if (_.size(data.members) < data.max_team_size) { %>
-            <p>You may have up to <%- data.max_team_size - _.size(data.members) %> more member<%- (data.max_team_size - _.size(data.members)) == 1 ? "" : "s" %> on your team.</p>
-            <p>New team members can join at any time.</p>
-        <% } else { %>
-            <p>Your team is full. No more members may join your team.</p>
-        <% }; %>
-        <% if (data.eligible) { %>
-            <p>Your team is <b>eligible</b>.</p>
-        <% } else { %>
-            <p>Your team is <b>ineligible</b> (<a href="#" class="poptrigger" id='eligibility-exp' data-container="body" data-toggle="popover" data-placement="right"
-                                                  data-content="">See Explanation</a>).</p>
-        <% } %>
-    </div>
   </div>
-</script>
-
-<script type="text/template" id="achievement-info-template">
-  <div class="panel panel-default">
-    <div class="panel-heading">
-      <h3 class="panel-title">Achievements</h3>
-    </div>
-    <div class="panel-body">
-        <div style="height: 300px; overflow: auto;">
-            <table>
-                <% if (_.size(data) > 0) { %>
-                    <% _.each(data, function(achievement) { %>
-                        <tr><td><img class="image-achievement" src='<%- achievement.image %>'/></td>
-                            <td style="padding-left:15px">
-                                <div><h3 style="margin-top:0px;"><%- achievement.name %></h3></div>
-                                <div><%- achievement.description %></div>
-                            </td>
-                        </tr>
-                    <% }); %>
-                <% } else { %>
-                    <tr><td>You currently have no achievements. Achievements are earned by solving challenges and playing through the story.</td></tr>
-                <% } %>
-            </table>
-        </div>
-    </div>
-  </div>
-</script>
-
-<script language="javascript">
-    $(() => {
-        $('.poptrigger').popover({html: true});
-    });
 </script>
