+++ picoCTF-web/web/reset.html
@@ -2,7 +2,7 @@
 layout: default
 title: Password Reset
 post_scripts:
- - js/account.js
+ - js/profile.js
 ---
 
 <div class="container">
